version 0.92

+++ QuickBooks/Models/Business/ReportService.cs
@@ -3,23 +3,23 @@ using Common.Logging;
 using QuickBooks.Models.DAL;
 using QuickBooks.Models.Repository;
 
-namespace QuickBooks.Models.ReportService
+namespace QuickBooks.Models.Business
 {
     public class ReportService : IReportService
     {
-        private readonly IReportRepository _repository;
-        private readonly ILog _log = LogManager.GetLogger<ReportService>();
-
         public ReportService(IReportRepository repository)
         {
             _repository = repository;
         }
+        private readonly IReportRepository _repository;
+        private readonly ILog _log = LogManager.GetLogger<ReportService>();
+        private string _entityId = "";
 
         public void Save(BaseEntity entity)
         {
             try
             {
-                var report = new Report()
+                var report = new Report
                 {
                     Id = entity.Id,
                     ShipToAddress = entity.ShipToAddress,
@@ -28,11 +28,12 @@ namespace QuickBooks.Models.ReportService
                     CustomerName = entity.CustomerName,
                     SaleDate = entity.SaleDate
                 };
-           _repository.Create(report);
+                _entityId = entity.Id;
+                _repository.Create(report);
             }
             catch (Exception e)
             {
-                _log.Error("Exception occured when you tried to save Report entity", e);
+                _log.Error($"Exception occured when application tried to save Report entity with id ={_entityId}", e);
                 throw;
             }
         }
@@ -41,19 +42,29 @@ namespace QuickBooks.Models.ReportService
         {
             try
             {
+                _entityId = id;
                 var report = _repository.Get(id);
                 return report;
             }
             catch (Exception e)
             {
-                _log.Error("Exception occured when you tried to find report by entityId", e);
+                _log.Error($"Exception occured when application tried to find report with id ={_entityId}", e);
                 throw;
             }
         }
 
         public void Delete(string id)
         {
-            _repository.Delete(id);
+            try
+            {
+                _entityId = id;
+                var entity = Get(id);
+                if (entity!=null) _repository.Delete(id);
+            }
+            catch (Exception e)
+            {
+                _log.Error($"Exception occured when application tried to delete report with id ={_entityId}", e);
+            }
         }
     }
 }
